Moved method related to users into app.business.users
